chore(virtctl): Move migrate cancel command to own file
