+++ pkg/virtctl/vm/vm.go
@@ -38,11 +38,10 @@ import (
 )
 
 const (
-	COMMAND_MIGRATE_CANCEL = "migrate-cancel"
-	COMMAND_GUESTOSINFO    = "guestosinfo"
-	COMMAND_USERLIST       = "userlist"
-	COMMAND_FSLIST         = "fslist"
-	COMMAND_ADDVOLUME      = "addvolume"
+	COMMAND_GUESTOSINFO = "guestosinfo"
+	COMMAND_USERLIST    = "userlist"
+	COMMAND_FSLIST      = "fslist"
+	COMMAND_ADDVOLUME   = "addvolume"
 
 	volumeNameArg         = "volume-name"
 	notDefinedGracePeriod = -1
@@ -69,21 +68,6 @@ var (
 	dryRun       bool
 	cache        string
 )
-
-func NewMigrateCancelCommand(clientConfig clientcmd.ClientConfig) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:     "migrate-cancel (VM)",
-		Short:   "Cancel migration of a virtual machine.",
-		Example: usage(COMMAND_MIGRATE_CANCEL),
-		Args:    templates.ExactArgs("migrate-cancel", 1),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			c := Command{command: COMMAND_MIGRATE_CANCEL, clientConfig: clientConfig}
-			return c.Run(args)
-		},
-	}
-	cmd.SetUsageTemplate(templates.UsageTemplate())
-	return cmd
-}
 
 func NewGuestOsInfoCommand(clientConfig clientcmd.ClientConfig) *cobra.Command {
 	cmd := &cobra.Command{
@@ -286,35 +270,6 @@ func (o *Command) Run(args []string) error {
 		fmt.Printf("Dry Run execution\n")
 	}
 	switch o.command {
-	case COMMAND_MIGRATE_CANCEL:
-		// get a list of migrations for vmiName (use LabelSelector filter)
-		labelselector := fmt.Sprintf("%s==%s", v1.MigrationSelectorLabel, vmiName)
-		migrations, err := virtClient.VirtualMachineInstanceMigration(namespace).List(&metav1.ListOptions{
-			LabelSelector: labelselector})
-		if err != nil {
-			return fmt.Errorf("Error fetching virtual machine instance migration list  %v", err)
-		}
-
-		// There may be a single active migrations but several completed/failed ones
-		// go over the migrations list and find the active one
-		done := false
-		for _, mig := range migrations.Items {
-			migname := mig.ObjectMeta.Name
-
-			if !mig.IsFinal() {
-				// Cancel the active migration by calling Delete
-				err = virtClient.VirtualMachineInstanceMigration(namespace).Delete(migname, &metav1.DeleteOptions{})
-				if err != nil {
-					return fmt.Errorf("Error canceling migration %s of a VirtualMachine %s: %v", migname, vmiName, err)
-				}
-				done = true
-				break
-			}
-		}
-
-		if !done {
-			return fmt.Errorf("Found no migration to cancel for %s", vmiName)
-		}
 	case COMMAND_GUESTOSINFO:
 		guestosinfo, err := virtClient.VirtualMachineInstance(namespace).GuestOsInfo(context.Background(), vmiName)
 		if err != nil {
